moved code useful for impl. of data cache from icache to seperate module coreblocks.cache
