+++ coreblocks/core.py
@@ -18,8 +18,9 @@ from coreblocks.structs_common.exception import ExceptionCauseRegister
 from coreblocks.scheduler.scheduler import Scheduler
 from coreblocks.stages.backend import ResultAnnouncement
 from coreblocks.stages.retirement import Retirement
-from coreblocks.frontend.icache import ICache, SimpleWBCacheRefiller, ICacheBypass
+from coreblocks.frontend.icache import ICache, ICacheBypass
 from coreblocks.peripherals.wishbone import WishboneMaster, WishboneBus
+from coreblocks.cache.refiller import SimpleWBCacheRefiller
 from coreblocks.frontend.fetch import Fetch, UnalignedFetch
 from transactron.lib.transformers import MethodMap, MethodProduct
 from transactron.utils.fifo import BasicFifo
