+++ test/frontend/test_icache.py
@@ -7,10 +7,12 @@ from amaranth.sim import Passive, Settle
 from amaranth.utils import log2_int
 
 from transactron.lib import AdapterTrans, Adapter
-from coreblocks.frontend.icache import SimpleWBCacheRefiller, ICache, ICacheBypass, CacheRefillerInterface
+from coreblocks.frontend.icache import ICache, ICacheBypass
 from coreblocks.params import GenParams, ICacheLayouts
 from coreblocks.peripherals.wishbone import WishboneMaster, WishboneParameters
 from coreblocks.params.configurations import test_core_config
+from coreblocks.cache.iface import CacheRefillerInterface
+from coreblocks.cache.refiller import SimpleWBCacheRefiller
 
 from ..common import TestCaseWithSimulator, TestbenchIO, def_method_mock, RecordIntDictRet
 from ..peripherals.test_wishbone import WishboneInterfaceWrapper
